Fixed ActionBar attribute
I had originally used ActionGroup to separate different sections of the ActionBar, primarily for aesthetics, but that was incorrect. Action group is a function of itself which acts like a dropdown when the screen is a certain size. I instead used ActionSeparator which will allow the division of groups without actually making an ActionGroup